third_party/blink/web_tests/external/wpt/css/css-fonts/parsing/font-size-adjust-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['font-size-adjust'] = "cap-height calc(10 + (sign(20cqw - 10px) * 5))" should set the property value
  assert_equals: serialization should be canonical expected "cap-height calc(10 + (5 * sign(20cqw - 10px)))" but got "cap-height calc(10 + sign(20cqw - 10px) * 5)"
Harness: the test ran to completion.
